reorganizing and form
I re-organized the folder structure and I add forms. I will finish it tomorrow.
