+++ public/Scripts/app.js
@@ -1,0 +1,11 @@
+(function(){
+
+    function Start()
+    {
+        console.log("App Started...");
+    
+    }
+
+    window.addEventListener("load",Start);
+
+})();
