+++ server/views/createsurvey/list.ejs
@@ -1,0 +1,7 @@
+<!-- Bohyun Kim, #301131832, COMP 229, Section 008 -->
+
+<%- include ../partials/header.ejs %>
+<%- include ../partials/main_nav.ejs %>
+
+<%- include ../partials/bottom_nav.ejs %>
+<%- include ../partials/footer.ejs %>
