+++ server/views/index.ejs
@@ -3,6 +3,13 @@
 <%- include partials/header.ejs %>
 <%- include partials/main_nav.ejs %>
 
-    <h4>General Human Data Collection Survey</h4>
+  <h2>Steller Surveys</h2>
+  <hr class="my-4">
+
+ 
+  <a href="/survey-list/add" class="btn btn-primary" role="button">
+    <i class = "fas fa-plus-circle"></i> Create Survey</a>
+</div>
+
 <%- include partials/bottom_nav.ejs %>
 <%- include partials/footer.ejs %>
