+++ server/views/partials/main_nav.ejs
@@ -20,9 +20,9 @@
 
           <li class="nav-item">
           <% if(title == "Survey") { %>
-            <a class="nav-link active" href="/survey"><i class="fas fa-lg fa-info"></i> Survey</a>
+            <a class="nav-link active" href="/survey-list"><i class="fas fa-lg fa-info"></i> Survey</a>
           <% } else { %>
-            <a class="nav-link" href="/survey"><i class="fas fa-lg fa-info"></i> About Me</a>
+            <a class="nav-link" href="/survey-list"><i class="fas fa-lg fa-info"></i> Survey</a>
           <% } %>
           </li>
