--- views/index.ejs
@@ -1,11 +1,0 @@
-<!--Akshit Narang, #301177669, COMP 229, Section 008-->
-
-<%- include partials/header.ejs %>
-<%- include partials/main_nav.ejs %>
-
-  <h2>Steller Surveys</h2>
-  <hr class="my-4">
-</div>
-
-<%- include partials/bottom_nav.ejs %>
-<%- include partials/footer.ejs %>
